Cleanup before merge
