+++ src/main/java/nl/slindenau/openmusicscrobbler/Application.java
@@ -2,6 +2,7 @@ package nl.slindenau.openmusicscrobbler;
 
 import nl.slindenau.openmusicscrobbler.cli.ConsoleClient;
 import nl.slindenau.openmusicscrobbler.cli.EncryptPasswordClient;
+import nl.slindenau.openmusicscrobbler.web.WebApplication;
 
 /**
  * @author slindenau
@@ -12,6 +13,7 @@ public class Application {
 
     private static final String CONSOLE_APPLICATION = "console";
     private static final String ENCRYPT_APPLICATION = "encrypt";
+    private static final String SERVER_APPLICATION = "server";
     private static final String DEFAULT_APPLICATION_TO_RUN = CONSOLE_APPLICATION;
 
     public static void main(String[] commandLineArguments) {
@@ -19,6 +21,7 @@ public class Application {
         switch (applicationToRun) {
             case ENCRYPT_APPLICATION -> new EncryptPasswordClient().run();
             case CONSOLE_APPLICATION -> new ConsoleClient().run();
+            case SERVER_APPLICATION -> new WebApplication().run(commandLineArguments);
             default -> throw new IllegalArgumentException("Unknown command: " + applicationToRun);
         }
     }
