+++ src/main/java/nl/slindenau/openmusicscrobbler/web/api/CollectionResource.java
@@ -1,6 +1,7 @@
-package nl.slindenau.openmusicscrobbler.web;
+package nl.slindenau.openmusicscrobbler.web.api;
 
 import com.codahale.metrics.annotation.Timed;
+import nl.slindenau.openmusicscrobbler.config.ApplicationProperties;
 import nl.slindenau.openmusicscrobbler.model.ReleaseCollection;
 import nl.slindenau.openmusicscrobbler.service.DiscogsService;
 import nl.slindenau.openmusicscrobbler.web.view.ReleaseCollectionView;
@@ -13,25 +14,32 @@ import java.util.Optional;
 @Produces({MediaType.APPLICATION_JSON+";charset=UTF-8", MediaType.TEXT_HTML+";charset=UTF-8"})
 public class CollectionResource {
 
-    //todo ding thread safe maken
+    private final ApplicationProperties applicationProperties = new ApplicationProperties();
+
+    //todo: make thread safe
     DiscogsService discogsService = new DiscogsService();
 
     @GET
     @Timed
     @Consumes({MediaType.APPLICATION_JSON})
     public ReleaseCollection getCollectionAsJson() {
-        return discogsService.getUserCollection("slindenau2");
+        return discogsService.getUserCollection(getDiscogsUsername());
     }
 
     @GET
     @Timed
     public ReleaseCollectionView getCollectionAsView() {
-        return new ReleaseCollectionView(discogsService.getUserCollection("slindenau2"));
+        return new ReleaseCollectionView(discogsService.getUserCollection(getDiscogsUsername()));
     }
 
     @POST
     @Path("/scrobble")
     public void scrobble(@QueryParam("id") Optional<Long> id) {
+        // todo: implement
         System.out.println("Scrobble: " + id.get());
+    }
+
+    private String getDiscogsUsername() {
+        return applicationProperties.getDiscogsUsername();
     }
 }
